server/requirements.txt
fastapi>=0.110.0
openenv-core>=0.2.0
openai>=1.30.0
pydantic>=2.6.0
PyYAML>=6.0.1
requests>=2.31.0
uvicorn[standard]>=0.29.0
